minor tweaks

+++ src/Game/IncomeTaxSquare.java
@@ -15,6 +15,6 @@ public class IncomeTaxSquare extends Square {
     @Override
     public void landedOn(Player p) {
         int w= p.getNetWorth();
-        p.reduceCash(Math.min(w*0.1f,200));
+        p.reduceCash(Math.min((int) (w*0.1f),200));
     }
 }
